requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.5
typing-extensions>=4.9
python-dotenv>=1.0
